skip compiling knit-tikzDevice.Rnw if pdflatex is not available

+++ tests/testit/test-plot.R
@@ -130,7 +130,7 @@ assert('plots are rearrange based on fig.keep & fig.show options', {
 })
 
 # should not error when a plot label contains special characters and sanitize=TRUE
-if (xfun::loadable('tikzDevice') &&
+if (xfun::loadable('tikzDevice') && Sys.which('pdflatex') != '' &&
     (!is.na(Sys.getenv('CI', NA)) || Sys.getenv('USER') == 'yihui' || !xfun::is_macos())) {
   knit('knit-tikzDevice.Rnw', quiet = TRUE)
   unlink(c('*-tikzDictionary', 'figure', 'knit-tikzDevice.tex'), recursive = TRUE)
